requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.3.0=pypi_0
accelerate=0.15.0=pypi_0
aiofiles=23.1.0=pypi_0
aiohttp=3.8.3=pypi_0
aiosignal=1.3.1=pypi_0
altair=4.2.2=pypi_0
anyio=3.6.2=pypi_0
async-timeout=4.0.2=pypi_0
attrs=22.1.0=pypi_0
blessed=1.19.1=pypi_0
blis=0.7.9=pypi_0
blobfile=2.0.0=pypi_0
ca-certificates=2022.10.11=h06a4308_0
cachetools=5.2.0=pypi_0
catalogue=2.0.8=pypi_0
certifi=2022.9.24=py38h06a4308_0
charset-normalizer=2.1.1=pypi_0
click=8.1.3=pypi_0
clip=1.0=pypi_0
commonmark=0.9.1=pypi_0
composable-diffusion=2.0=pypi_0
confection=0.0.3=pypi_0
contourpy=1.0.5=pypi_0
cycler=0.11.0=pypi_0
cymem=2.0.7=pypi_0
daam=0.0.12=pypi_0
datasets=2.7.1=pypi_0
dearpygui=1.7.1=pypi_0
diffuser=0.0.1=pypi_0
diffusers=0.10.2=pypi_0
dill=0.3.6=pypi_0
easydict=1.10=pypi_0
entrypoints=0.4=pypi_0
fastapi=0.92.0=pypi_0
ffmpy=0.3.0=pypi_0
filelock=3.8.0=pypi_0
fonttools=4.37.4=pypi_0
frozenlist=1.3.3=pypi_0
fsspec=2022.11.0=pypi_0
ftfy=6.1.1=pypi_0
google-auth=2.14.1=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
gpustat=1.0.0=pypi_0
gpustats=0.0.1=pypi_0
gradio=3.19.1=pypi_0
grpcio=1.50.0=pypi_0
h11=0.14.0=pypi_0
httpcore=0.16.3=pypi_0
httpx=0.23.3=pypi_0
huggingface-hub=0.14.1=pypi_0
idna=3.4=pypi_0
imageio=2.25.1=pypi_0
imageio-ffmpeg=0.4.8=pypi_0
importlib-metadata=5.0.0=pypi_0
importlib-resources=5.12.0=pypi_0
inflect=6.0.2=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
jsonschema=4.17.3=pypi_0
kiwisolver=1.4.4=pypi_0
langcodes=3.3.0=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
linkify-it-py=2.0.0=pypi_0
llvmlite=0.39.1=pypi_0
lpips=0.1.4=pypi_0
lxml=4.9.2=pypi_0
markdown=3.4.1=pypi_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.6.0=pypi_0
mdit-py-plugins=0.3.3=pypi_0
mdurl=0.1.2=pypi_0
modelcards=0.1.6=pypi_0
multidict=6.0.2=pypi_0
multiprocess=0.70.14=pypi_0
murmurhash=1.0.9=pypi_0
ncurses=6.3=h5eee18b_3
networkx=3.0=pypi_0
ninja=1.10.2.4=pypi_0
nltk=3.8.1=pypi_0
numba=0.56.4=pypi_0
numpy=1.23.3=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-ml-py=11.495.46=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.7.0.72=pypi_0
openssl=1.1.1s=h7f8727e_0
orjson=3.8.6=pypi_0
packaging=21.3=pypi_0
pandas=1.5.0=pypi_0
pathy=0.10.0=pypi_0
pillow=9.3.0=pypi_0
pip=22.3.1=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
preshed=3.0.8=pypi_0
protobuf=3.20.1=pypi_0
psutil=5.9.2=pypi_0
pyarrow=10.0.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.6=pypi_0
pycryptodome=3.17=pypi_0
pycryptodomex=3.16.0=pypi_0
pydantic=1.10.2=pypi_0
pydub=0.25.1=pypi_0
pygments=2.13.0=pypi_0
pymcubes=0.1.2=pypi_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.19.3=pypi_0
python=3.8.15=h7a1cb2a_2
python-dateutil=2.8.2=pypi_0
python-multipart=0.0.5=pypi_0
pytz=2022.4=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
readline=8.2=h5eee18b_0
regex=2022.9.13=pypi_0
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
responses=0.18.0=pypi_0
rfc3986=1.5.0=pypi_0
rsa=4.9=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.1.2=pypi_0
scipy=1.9.3=pypi_0
setuptools=59.5.0=pypi_0
six=1.16.0=pypi_0
smart-open=5.2.1=pypi_0
sniffio=1.3.0=pypi_0
spacy=3.4.3=pypi_0
spacy-legacy=3.0.10=pypi_0
spacy-loggers=1.0.3=pypi_0
sqlite=3.40.0=h5082296_0
srsly=2.4.5=pypi_0
starlette=0.25.0=pypi_0
tensorboard=2.11.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorboardx=2.5.1=pypi_0
thinc=8.1.5=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2023.2.3=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.12.1=pypi_0
toolz=0.12.0=pypi_0
torch=1.11.0+cu113=pypi_0
torchaudio=0.11.0+cu113=pypi_0
torchmetrics=0.11.1=pypi_0
torchvision=0.12.0+cu113=pypi_0
tqdm=4.64.1=pypi_0
transformers=4.26.0=pypi_0
trimesh=3.15.3=pypi_0
typer=0.7.0=pypi_0
typing-extensions=4.3.0=pypi_0
uc-micro-py=1.0.1=pypi_0
urllib3=1.26.13=pypi_0
uvicorn=0.20.0=pypi_0
wasabi=0.10.1=pypi_0
wcwidth=0.2.5=pypi_0
websockets=10.4=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xatlas=0.0.6=pypi_0
xxhash=3.1.0=pypi_0
xz=5.2.8=h5eee18b_0
yarl=1.8.1=pypi_0
zipp=3.8.1=pypi_0
zlib=1.2.13=h5eee18b_0
